Add infrastructure to retain comments matching a wildcard pattern.  This still needs to be exposed as an Apache directive and doc'd.

+++ src/net/instaweb/rewriter/remove_comments_filter.cc
@@ -19,14 +19,15 @@
 #include "net/instaweb/rewriter/public/remove_comments_filter.h"
 
 #include "net/instaweb/htmlparse/public/html_parse.h"
+#include "net/instaweb/rewriter/public/rewrite_options.h"
 
 namespace net_instaweb {
 
-RemoveCommentsFilter::RemoveCommentsFilter(HtmlParse* html_parse)
-    : html_parse_(html_parse) {}
-
 void RemoveCommentsFilter::Comment(HtmlCommentNode* comment) {
-  html_parse_->DeleteElement(comment);
+  if ((rewrite_options_ == NULL) ||
+      !rewrite_options_->IsRetainedComment(comment->contents())) {
+    html_parse_->DeleteElement(comment);
+  }
 }
 
 }  // namespace net_instaweb
